Generic macro expansion support
Required for CBG-3291.  Refactors MutateInSpec to MacroExpansionSpec, to be used specifically for defining the set of macro expansions that should be performed for a given xattr operation.

Added MutateInOptions to XattrStore methods that need to specify this.  Removed a few XattrStore methods that are only used internally in the gocb implementation and don't need to be on the interface.